requirements.txt
asgiref==3.11.1
bcrypt==4.3.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.4.0
PyMySQL==1.1.1
pycryptodome==3.23.0
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
django-ipware==7.0.1